@@ -3,13 +3,14 @@
 	"description": "Web interface for Intuition.",
 	"license": "MIT",
 	"require": {
-		"php": ">=7.2",
-		"krinkle/toollabs-base": "^1.0.2",
-		"krinkle/intuition": "^2.3.1"
+		"php": ">=7.4",
+		"krinkle/toollabs-base": "^2.1.1",
+		"krinkle/intuition": "^2.3.4"
 	},
 	"require-dev": {
-		"jakub-onderka/php-parallel-lint": "1.0.0",
-		"mediawiki/mediawiki-codesniffer": "27.0.0"
+		"php-parallel-lint/php-parallel-lint": "1.3.2",
+		"mediawiki/mediawiki-codesniffer": "40.0.1",
+		"mediawiki/mediawiki-phan-config": "0.12.0"
 	},
 	"config": {
 		"process-timeout": 0
@@ -18,7 +19,10 @@
 		"serve": "php -S localhost:4000 -t public_html/",
 		"test": [
 			"parallel-lint . --exclude libs --exclude vendor --exclude node_modules",
+			"@phan",
 			"phpcs -p -s"
-		]
+		],
+		"phan": "phan --allow-polyfill-parser --no-progress-bar",
+		"fix": "phpcbf"
 	}
 }
